Fix issues flagged by PVS Studio (#7839)
* Fix issues flagged by PVS Studio

* Fixes

+++ src/Orleans.Runtime/Cancellation/CancellationSourcesExtension.cs
@@ -127,7 +127,7 @@ namespace Orleans.Runtime
 
             public bool IsExpired(TimeSpan expiry, long nowTimestamp)
             {
-                var untouchedTime = TimeSpan.FromSeconds((nowTimestamp - _createdTime) / Stopwatch.Frequency);
+                var untouchedTime = TimeSpan.FromSeconds((nowTimestamp - _createdTime) / (double)Stopwatch.Frequency);
 
                 return untouchedTime >= expiry;
             }
